requirements.txt
playwright==1.44.0
openai-whisper
groq
anthropic==0.28.0
fastapi==0.111.0
uvicorn==0.30.1
apscheduler==3.10.4
sqlalchemy==2.0.30
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
jinja2==3.1.4
